replayer: Make LLVMFuzzerInitialize optional on macOS and Windows
The macOS linker requires an explicit -U <symbol> flag if a weak symbol
is not overriden, so we use a dlsym lookup instead, which works on macOS
without additional flags. It would require -rdynamic or --export-dynamic
on Linux and thus can't be used there.

The Windows linker cannot deal with weak symbols combined with exception
infos (which are generated by default), so we /alternatename (MSVC) and
alias (clang) instead.

These approaches are based on how libFuzzer handles optional functions.