cmake: Set compiler and linker options
Add the same compiler and linker options as are set for make
invocations when using clang or gcc and adapt the set of flags for
MSVC-style compilers.

As we don't want to impose fuzzing-related flags on the entire build, we
gate these flags behind a CIFUZZ_TESTING cache variable. In the future,
we could also explore whether this is better modelled as a build type.

The UBSan example needed a small tweak as with -O2 (set by
RelWithDebInfo), the compiler is smart enough to figure out that the
comparison triggering UBSan can only be satisfied if a signed integer
overflow occurs - which is UB. As a result, the compiler optimized out
the check, leading to an integration test failure.