Return ErrSilent instead of wrapping the original error
Returning the wrapped original error implies that it might still be used
for something (e.g. to print the error message), which is not the case.
Returning an ErrSilent without any information about the original error
makes it clear to the caller that they have to print the error message
themselves.