Fix wrapping IncorrectUsageError
We used to wrap errors into incorrect usage errors like this:

    fmt.Errorf("%w:\n%+v", ErrIncorrectUsage, err)

This made the error message contain "IncorrectUsageError:", which we
don't want to show the user, *and* it was not possible to unwrap the
error into the original error, because what was being wrapped via "%w"
was the ErrIncorrectUsage, not the original error.

We now do the same gh does with their FlagError, which is to use a
struct which implements Unwrap() to unwrap the IncorrectUsageError
into the original error. Note that this requires using `errors.As`
instead of `errors.Is`.