runner: Print startup stderr in case of an unexpected exit error
Without the stderr of the fuzzer printed during startup, it isn't
possible for users to understand why it exits abnormally, e.g. due to a
missing shared library dependency.

With this commit, the parsed output up to the point where the fuzzer has
been initialized is stored and made available to the runner to print in
case of an error.