Call fsync after copying
We've seen some errors in CI pipelines indicating that the files copied
by the installer are being used before they were synced to disk, so we
now sync after copying.

The Copy function from github.com/otiai10/copy already supports syncing
after copying, so we're using that now instead of fileutil.CopyFile.

Note that this might not fix all synchronization issues, quoting the
fsync(2) manual:

    Calling fsync() does not necessarily ensure that the entry in the directory
    containing the file has also reached disk.  For that an explicit fsync() on
    a file descriptor for the directory is also needed.

So in case we still see synchronization issues, we should also call fsync
on the directory of the file we copied.