cmake: Ensure that replayer is compiled into CXX-only projects
The fix in 9d3ecb096d4f2cc80c363c375e0d7bb88042baaa was incomplete as it
only modified the linker language, not the source file language. As a
result, in CXX-only projects, the replayer static library would be empty.

This is fixed by marking the source file as C++ so that CMake can use
the configured C++ compiler, but passing in flags that make the C++
compiler treat the file as C. Just in case, the previous commit ensures
that it compiles just fine treated as C++.